Move replica rebalancing to ClusterBalancer
