+++ kafka_utils/kafka_cluster_manager/cluster_info/cluster_balancer.py
@@ -12,6 +12,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import logging
+
+from .util import separate_groups
 
 
 class ClusterBalancer(object):
@@ -24,6 +27,7 @@ class ClusterBalancer(object):
     def __init__(self, cluster_topology, args):
         self.cluster_topology = cluster_topology
         self.args = args
+        self.log = logging.getLogger(self.__class__.__name__)
 
     def rebalance(self):
         """Rebalance partitions across the brokers in the cluster."""
@@ -57,3 +61,80 @@ class ClusterBalancer(object):
         :raises InvalidReplicationFactorError: The resulting replication factor is invalid.
         """
         raise NotImplementedError("Implement in subclass")
+
+    def rebalance_replicas(self):
+        """Balance replicas across replication-groups."""
+        for partition in self.cluster_topology.partitions.itervalues():
+            self._rebalance_partition_replicas(partition)
+
+    def _rebalance_partition_replicas(self, partition):
+        """Rebalance replication group for given partition."""
+        # Separate replication-groups into under and over replicated
+        total = partition.replication_factor
+        over_replicated_rgs, under_replicated_rgs = separate_groups(
+            self.cluster_topology.rgs.values(),
+            lambda g: g.count_replica(partition),
+            total,
+        )
+        # Move replicas from over-replicated to under-replicated groups
+        while under_replicated_rgs and over_replicated_rgs:
+            # Decide source and destination group
+            rg_source = self._elect_source_replication_group(
+                over_replicated_rgs,
+                partition,
+            )
+            rg_destination = self._elect_dest_replication_group(
+                rg_source.count_replica(partition),
+                under_replicated_rgs,
+                partition,
+            )
+            if rg_source and rg_destination:
+                # Actual movement of partition
+                self.log.debug(
+                    'Moving partition {p_name} from replication-group '
+                    '{rg_source} to replication-group {rg_dest}'.format(
+                        p_name=partition.name,
+                        rg_source=rg_source.id,
+                        rg_dest=rg_destination.id,
+                    ),
+                )
+                rg_source.move_partition(rg_destination, partition)
+            else:
+                # Groups balanced or cannot be balanced further
+                break
+            # Re-compute under and over-replicated replication-groups
+            over_replicated_rgs, under_replicated_rgs = separate_groups(
+                self.cluster_topology.rgs.values(),
+                lambda g: g.count_replica(partition),
+                total,
+            )
+
+    def _elect_source_replication_group(
+            self,
+            over_replicated_rgs,
+            partition,
+    ):
+        """Decide source replication-group based as group with highest replica
+        count.
+        """
+        return max(
+            over_replicated_rgs,
+            key=lambda rg: rg.count_replica(partition),
+        )
+
+    def _elect_dest_replication_group(
+            self,
+            replica_count_source,
+            under_replicated_rgs,
+            partition,
+    ):
+        """Decide destination replication-group based on replica-count."""
+        min_replicated_rg = min(
+            under_replicated_rgs,
+            key=lambda rg: rg.count_replica(partition),
+        )
+        # Locate under-replicated replication-group with lesser
+        # replica count than source replication-group
+        if min_replicated_rg.count_replica(partition) < replica_count_source - 1:
+            return min_replicated_rg
+        return None
