+++ kafka_utils/kafka_cluster_manager/cluster_info/partition_count_balancer.py
@@ -159,8 +159,8 @@ class PartitionCountBalancer(ClusterBalancer):
                 "brokers"
             )
 
-        for partition in self.cluster_topology.partitions.itervalues():
-            self._rebalance_partition(partition)
+        # Balance partition-count over replication-groups
+        self.rebalance_replicas()
 
         # Balance partition-count over replication-groups
         self._rebalance_groups_partition_cnt()
@@ -221,81 +221,6 @@ class PartitionCountBalancer(ClusterBalancer):
             for broker in over_brokers:
                 skip_brokers.append(broker)
                 broker.donate_leadership(opt_cnt, skip_brokers, used_edges)
-
-    def _rebalance_partition(self, partition):
-        """Rebalance replication group for given partition."""
-        # Separate replication-groups into under and over replicated
-        total = sum(
-            group.count_replica(partition)
-            for group in self.cluster_topology.rgs.itervalues()
-        )
-        over_replicated_rgs, under_replicated_rgs = separate_groups(
-            self.cluster_topology.rgs.values(),
-            lambda g: g.count_replica(partition),
-            total,
-        )
-        # Move replicas from over-replicated to under-replicated groups
-        while under_replicated_rgs and over_replicated_rgs:
-            # Decide source and destination group
-            rg_source = self._elect_source_replication_group(
-                over_replicated_rgs,
-                partition,
-            )
-            rg_destination = self._elect_dest_replication_group(
-                rg_source.count_replica(partition),
-                under_replicated_rgs,
-                partition,
-            )
-            if rg_source and rg_destination:
-                # Actual movement of partition
-                self.log.debug(
-                    'Moving partition {p_name} from replication-group '
-                    '{rg_source} to replication-group {rg_dest}'.format(
-                        p_name=partition.name,
-                        rg_source=rg_source.id,
-                        rg_dest=rg_destination.id,
-                    ),
-                )
-                rg_source.move_partition(rg_destination, partition)
-            else:
-                # Groups balanced or cannot be balanced further
-                break
-            # Re-compute under and over-replicated replication-groups
-            over_replicated_rgs, under_replicated_rgs = separate_groups(
-                self.cluster_topology.rgs.values(),
-                lambda g: g.count_replica(partition),
-                total,
-            )
-
-    def _elect_source_replication_group(
-        self,
-        over_replicated_rgs,
-        partition,
-    ):
-        """Decide source replication-group based as group with highest replica
-        count.
-        """
-        return max(
-            over_replicated_rgs,
-            key=lambda rg: rg.count_replica(partition),
-        )
-
-    def _elect_dest_replication_group(
-        self,
-        replica_count_source,
-        under_replicated_rgs,
-        partition,
-    ):
-        """Decide destination replication-group based on replica-count."""
-        min_replicated_rg = min(
-            under_replicated_rgs,
-            key=lambda rg: rg.count_replica(partition),
-        )
-        # Locate under-replicated replication-group with lesser
-        # replica count than source replication-group
-        if min_replicated_rg.count_replica(partition) < replica_count_source - 1:
-            return min_replicated_rg
-        return None
 
     # Re-balancing partition count across brokers
     def _rebalance_groups_partition_cnt(self):
